--- src/recoil/atoms/testState.js
@@ -1,8 +1,0 @@
-import {atom} from 'recoil';
-
-export const testState = atom({
-    key: 'testState',
-    default: {
-      name: 'Test'
-    }
-});
