+++ src/recoil/atoms/testState.ts
@@ -1,0 +1,8 @@
+import { atom } from 'recoil';
+
+export const testState = atom({
+  key: 'testState',
+  default: {
+    name: 'Test',
+  },
+});
